fix tests depending on other components' tests
