+++ src/Symfony/Component/Form/Test/FormPerformanceTestCase.php
@@ -11,8 +11,6 @@
 
 namespace Symfony\Component\Form\Test;
 
-use Symfony\Component\Form\Tests\VersionAwareTest;
-
 /**
  * Base class for performance tests.
  *
