+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/BaseTypeTest.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\Test\TypeTestCase;
-use Symfony\Component\Form\Tests\VersionAwareTest;
+use Symfony\Component\Form\Test\VersionAwareTest;
 
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
